+++ docs/make.jl
@@ -40,7 +40,7 @@ makedocs(;
         "Home" => "index.md",
         "Linesearch" => "linesearch.md",
         "Linesearches on Manifolds" => "linesearch_on_manifolds.md",
-        "Weight decay on a manifold" => "weight_decay.md",
+        "Weight Decay on Manifolds" => "weight_decay.md",
         "References" => "references.md",
     ],
 )